+++ packages/core/src/lib/helpers/detect-browser.ts
