+++ packages/core/src/lib/helpers/index.ts
@@ -1,3 +1,3 @@
 export * from "./waitFor";
 export * from "./getActiveAccount";
-export * from "./isBrowserSupported";
+export * from "./detect-browser";
